Patched screen size and added fixed Level for testing
I fixed patched the screen size (not perfect but functional), and made generateDesign make a long stretch of ground to run on with holes for reference. I also added a dispose function. I also didn't test it, but if you dispose then immediatley create a new level class also specifying the newEnd in the constructor (close to where the player is) it'll generate a more level for the player to run on.