Mario now moves
I fixed it so Mario would move by making him continually have a linear impluse applied upon being defined and everytime update is called. I also fixed a bug that was preventing from working where there was a class body variable and a local body variable that should not have been. I left it commed out with a comment to show it.